% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_league.R
\name{create_league_r}
\alias{create_league_r}
\title{Process for creating a new league}
\usage{
create_league_r(
  league_name = file.path(golem::get_golem_wd(), "user-data", "new-league"),
  file_name,
  add_user = F,
  csv_init = c("date_time", "home_team", "away_team", "home_pred", "away_pred",
    "submit_time")
)
}
\arguments{
\item{league_name, file_name, add_user, csv_init}{hsha}
}
\description{
Expects a csv file with two entries user_name and user_pw where user_pw is the unencrypted passwords.
This csv file with user_name and user_pw should not stay with the rest of the source code with the app. It is only used
temporarily.
1. Create folder structure
2. Initialize the user specific csv files with dates, team names etc. 
3. Creates a meta data file with usernames and encrypted passwords
create_league_r
}
