% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dev_mode.R
\name{mod_dev_mode_ui}
\alias{mod_dev_mode_ui}
\alias{mod_dev_mode_server}
\title{mod_dev_mode_ui and mod_dev_mode_server}
\usage{
mod_dev_mode_ui(id)

mod_dev_mode_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
