% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{This will run the hello world example app.
Type right here to document code}
\usage{
run_app(msg = "INSERT TITLE HERE")
}
\arguments{
\item{msg}{Title heading to put on first tab}
}
\value{
No return value, it just runs the app!
}
\description{
This will run the hello world example app.
Type right here to document code
}
\examples{
golemTest::run_app()
run_app()

}
