% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab1_server.R, R/mod_tab1_ui.R
\name{mod_tab1_server}
\alias{mod_tab1_server}
\alias{mod_tab1_ui}
\title{mod_tab1_server}
\usage{
mod_tab1_server(input, output, session)

mod_tab1_ui(id)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{id}{shiny id}
}
\description{
Tab1 server module

A shiny Module.
}
\keyword{internal}
